Move AsVector func.
